First Implementation on Core Data Working
SaveChat and loadChatsFromCore have been added. The first saves all the
chats that are currently present in the model, the second loads every
chat that has been saved with core data. Note that the first function
does not check if there already is a similar record, so will create
duplicates of the same chat, user and messages for the time being every
time it is called